TriggerAffection/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-cosmos
requests
beautifulsoup4
lxml
python-dateutil
selenium
webdriver-manager
tenacity
pyodbc
